+++ crates/cubek-std/src/tile/compute/softmax.rs
@@ -2,13 +2,42 @@ use cubecl;
 use cubecl::prelude::*;
 
 use crate::StageIdent;
-use crate::tile::compute::bounce::{cmma_to_whitebox_fragment, whitebox_fragment_to_cmma};
+use crate::tile::compute::copy::{cmma_to_whitebox_fragment, whitebox_fragment_to_cmma};
 use crate::tile::compute::mask::{Mask, MaskExpand};
 use crate::tile::compute::rowwise::reducer::{fragment_row_max, fragment_row_sum};
 use crate::tile::data::{
-    BounceTile, InnerLayout, RegisterTile, RowWise, UnitTile, WhiteboxFragment,
+    BounceTile, InnerLayout, RegisterTile, RowWise, RowWiseExpand, UnitTile, WhiteboxFragment,
 };
-use crate::tile::{LOGIT_MASKED, Plane, Tile, TileExpand};
+use crate::tile::{Plane, Tile, TileExpand};
+
+/// Logits below this are considered masked (effectively -inf).
+/// Value chosen to fit within f16 range (~-65,504 max).
+pub const LOGIT_MASKED: f32 = -6e4;
+
+/// Any value smaller than this is considered numerically zero (used for
+/// fully-masked rows or tiny contributions). Value chosen to be above f16
+/// smallest normal (~6.1e-5).
+pub const FULLY_MASKED_ROW_THRESHOLD: f32 = 1e-4;
+
+#[cube]
+impl<E: Float> RowWise<E> {
+    /// Replaces each value `v` (v >= 0) in a row with `1/v`.
+    ///
+    /// If `v = 0`, the result is set to `0` instead of `1/0`.
+    /// This occurs when the entire row is masked, meaning it should
+    /// contribute no information, and ensures numerical stability.
+    pub fn recip_inplace(&mut self) {
+        for i in 0..self.num_rows {
+            let row_val = self.vals[i];
+
+            let epsilon = E::new(FULLY_MASKED_ROW_THRESHOLD);
+            let not_masked = E::cast_from(row_val >= epsilon);
+            let safe_val = clamp_min(row_val, epsilon);
+            let recip = safe_val.recip();
+            self.vals[i] = not_masked * recip;
+        }
+    }
+}
 
 /// Comptime descriptor for the row-shape used by online softmax. Determines
 /// how many rows per unit each running-state vector holds.
